Refactor: dividido sidenav
