+++ src/components/SideNav/SideNav.tsx
@@ -1,0 +1,25 @@
+import * as React from 'react';
+
+import MyVisualizations from './MyVisualizations';
+import RecentVisualizations from './RecentVisualizations';
+import SharedVisualizations from './SharedVisualizations';
+import SocialMenu from './SocialMenu';
+
+import SideBar from '../style/SideBar';
+
+
+class SideNav extends React.Component {
+
+    public render() {
+        return (
+            <SideBar>
+                <SocialMenu />
+                <RecentVisualizations />
+                <MyVisualizations />
+                <SharedVisualizations />
+            </SideBar>
+        );
+    }
+}
+
+export default SideNav;
