+++ src/components/style/SideBar.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+export default function SideBar(props: React.Props<any>) {
+    return (
+        <nav className="col-md-2 d-none d-md-block bg-light sidebar">
+            <div className="sidebar-sticky">
+                {props.children}
+            </div>
+        </nav>
+    );
+}
