+++ src/components/style/SideBarGroupItem.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export default function SideBarGroupItem(props: React.Props<any>) {
+    return (
+        <li className="nav-item">
+            {props.children}
+        </li>
+    );
+}
