+++ src/components/style/SideBarHeading.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export default function SideBarHeading(props: React.Props<any>) {
+    return (
+        <h6 className="sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted">
+            {props.children}
+        </h6>
+    );
+}
